requirements.txt
asgiref==3.5.2
attrs==21.4.0
bcrypt==3.2.2
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==37.0.2
distro==1.7.0
Django==4.0.5
djangorestframework==3.13.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
idna==3.3
jsonschema==3.2.0
paramiko==2.11.0
pycparser==2.21
PyNaCl==1.5.0
pyrsistent==0.18.1
python-dotenv==0.20.0
pytz==2022.1
PyYAML==5.4.1
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
texttable==1.6.4
urllib3==1.26.9
websocket-client==0.59.0
